Refactor k-text
