+++ panel/src/ui/components/Text.vue
@@ -12,8 +12,20 @@
 <script>
 export default {
   props: {
+    /**
+     * Changes the text alignment.
+     * Available options: `left`|`center`|`right`
+     */
     align: String,
+    /**
+     * Changes the font size of the text.
+     * Available options: `tiny`|`small`|`medium`|`large`|`huge`
+     */
     size: String,
+    /**
+     * Changes the layout/design of the text.
+     * Available options: `help`
+     */
     theme: String
   }
 }
@@ -50,7 +62,7 @@ export default {
   font-size: $text-base;
 }
 .k-text[data-size="large"] {
-  font-size: $text-lg;
+  font-size: $text-xl;
 }
 
 .k-text[data-theme="help"] {
